+++ euchre/constants.py
@@ -1,0 +1,11 @@
+# Card creation constants
+VALUES = (9, 10, 11, 12, 13, 14)
+SUITS = ("Spades", "Diamonds", "Clubs", "Hearts")
+
+# General gameplay constants
+PLAYER_COUNT = 4
+TEAM_COUNT = 2
+MAX_CARD_HAND_LIMIT = 5
+POINTS_TO_WIN = 10
+
+BOTS = ["Cow", "Dog", "Cat", "Farmer"]
